adjust IdentityBlockTest
The test was accounting for state machine cancellations when testing whether the block is cancelled correctly. With FlowReduxStateMachineFactory the state machine is not cancelled during the test and instead will be torn down at the end of the test with backgroundScope of runTest. So the tests now don't need to account for the state machine itself anymore and can just check the actual cancellations they're interested in.