remove LegacyFlowReduxStateMachine (#897)
* remove LegacyFlowReduxStateMachine

* ktlint

* api dump

* adjust IdentityBlockTest

The test was accounting for state machine cancellations when testing whether the block is cancelled correctly. With FlowReduxStateMachineFactory the state machine is not cancelled during the test and instead will be torn down at the end of the test with backgroundScope of runTest. So the tests now don't need to account for the state machine itself anymore and can just check the actual cancellations they're interested in.

* use RENDEZVOUS channel in tests for more predictable results

* resolve test timing issue caused by scope change

The state machine is now running in a different scope because of the shareIn which causes a timing issue. This simply makes the test wait for the first action to be received by the handler before advancing time.

* make OnAction ordered test actual ordering and add unordered test

* ktlint

* reduce timeout for js because promise has a lower timeout